// packages/neutral/src/generated/accounts/bundle.ts
export type BundleArgs = {
  name: ReadonlyUint8Array;
  manager: Address;
  keeper: Address;
  treasuryAccount: Address;
  allocatedReceivers: Array<Address>;
  bundleUnderlyingBalance: number | bigint;
  maxDepositAmount: number | bigint;
  withdrawalDelay: number | bigint;
  performanceFee: number;
  managementFeeBps: number;
  depositFee: number;
  withdrawalFee: number;
  managerPfeeShares: number | bigint;
  currentAllocationBps: number;
  oracleBuffer: number | bigint;
  totalShares: number | bigint;
  assetPrecision: number | bigint;
  assetAddress: Address;
  assetDecimals: number;
  withdrawalTMin: number | bigint;
  withdrawalTMax: number | bigint;
  withdrawalCurve: number;
  permissionned: boolean;
  managerMfeeShares: number | bigint;
  minDepositAmount: number | bigint;
  oracleUpdateTimeLimit: number | bigint;
  oracleMaxAge: number | bigint;
  withdrawalRedemptionRequestCutoffTs: number | bigint;
  withdrawalRedemptionUnlockCurrentCycleTs: number | bigint;
  withdrawalRedemptionUnlockNextCycleTs: number | bigint;
  referrerEnabled: boolean;
  referrerMinDepositAmount: number | bigint;
  referralTiers: Array<ReferralTierArgs>;
  tierCount: number;
  padding: ReadonlyUint8Array;
};