// packages/neutral/src/generated/accounts/userBundleAccount.ts
export type UserBundleAccountArgs = {
  owner: Address;
  lastDepositTimestamp: number | bigint;
  shares: number | bigint;
  pendingDeposit: number | bigint;
  pendingShares: number | bigint;
  estimatedPendingWithdrawalValue: number | bigint;
  withdrawalAvailableTimestamp: number | bigint;
  lastWithdrawalProcessTimestamp: number | bigint;
  lastHighWaterMark: number | bigint;
  hwmPerShare: number | bigint;
  lastManagementFeeTimestamp: number | bigint;
  netDeposits: number | bigint;
  totalFeeCharged: number | bigint;
  customDepositFeeBps: number;
  customWithdrawalFeeBps: number;
  customPerformanceFeeBps: number;
  customManagementFeeBps: number;
  feeOverrideFlags: number;
  customWithdrawalDelay: number | bigint;
  customWithdrawalTMin: number | bigint;
  customWithdrawalTMax: number | bigint;
  customWithdrawalCurve: number;
  withdrawalTimingOverrideFlags: number;
  switchActive: boolean;
  switchTargetBundle: Address;
  switchCreatedAt: number | bigint;
  referrer: Address;
  padding: ReadonlyUint8Array;
};